refactor: split Media and Header into smaller files
