+++ app/src/components/Media/Detail/index.tsx
@@ -1,0 +1,2 @@
+// eslint-disable-next-line no-restricted-exports
+export { default } from './MediaDetail';
