+++ app/src/components/Media/Item/MediaItem.tsx
@@ -1,0 +1,19 @@
+import { IMedia } from '~/services/tmdb/tmdb.types';
+import BannerItem from './BannerItem';
+import CardItem from './CardItem';
+
+interface IMediaItem {
+  type: 'banner' | 'card';
+  item: IMedia;
+}
+
+const MediaItem = (props: IMediaItem) => {
+  const { type, item } = props;
+
+  if (type === 'banner') {
+    return <BannerItem item={item} />;
+  }
+  return <CardItem item={item} />;
+};
+
+export default MediaItem;
