+++ app/src/components/Media/List/MediaListTable.tsx
@@ -1,0 +1,36 @@
+import { Table } from '@nextui-org/react';
+import { Link } from '@remix-run/react';
+import { IMedia } from '~/services/tmdb/tmdb.types';
+
+const MediaListTable = ({ items }: { items: IMedia[] }) => (
+  <Table
+    bordered
+    striped
+    color="secondary"
+    aria-label="Example pagination table"
+    css={{
+      height: 'auto',
+      minWidth: '100%',
+    }}
+  >
+    <Table.Header>
+      <Table.Column>Title</Table.Column>
+      <Table.Column>Vote</Table.Column>
+    </Table.Header>
+    <Table.Body>
+      {items.map((item) => {
+        const href = (item.mediaType === 'movie' ? '/movies/' : '/tv-shows/') + item.id;
+        return (
+          <Table.Row key={item.id}>
+            <Table.Cell>
+              <Link to={href}>{item.title}</Link>
+            </Table.Cell>
+            <Table.Cell>{item.voteAverage}</Table.Cell>
+          </Table.Row>
+        );
+      })}
+    </Table.Body>
+  </Table>
+);
+
+export default MediaListTable;
